+++ Models/MainTbl.cs
@@ -12,10 +12,10 @@ namespace Fag_el_Gamous.Models
         public string BurialId { get; set; }
         public string BurialLocationNs { get; set; }
         public string BurialLocationEw { get; set; }
-        public long? LowPairNs { get; set; }
-        public long? HighPairNs { get; set; }
-        public long? LowPairEw { get; set; }
-        public long? HighPairEw { get; set; }
+        public int? LowPairNs { get; set; }
+        public int? HighPairNs { get; set; }
+        public int? LowPairEw { get; set; }
+        public int? HighPairEw { get; set; }
         public string BurialSubplot { get; set; }
         public double? BurialDepth { get; set; }
         public double? SouthToHead { get; set; }
@@ -23,20 +23,20 @@ namespace Fag_el_Gamous.Models
         public double? WestToHead { get; set; }
         public double? WestToFeet { get; set; }
         public string BurialSituation { get; set; }
-        public long? LengthOfRemains { get; set; }
-        public long? BurialNumber { get; set; }
-        public long? SampleNumber { get; set; }
+        public int? LengthOfRemains { get; set; }
+        public int? BurialNumber { get; set; }
+        public int? SampleNumber { get; set; }
         public string GenderGe { get; set; }
         public double? GeFunctionTotal { get; set; }
         public string GenderBodyCol { get; set; }
         public string BasilarSuture { get; set; }
-        public long? VentralArc { get; set; }
-        public long? SubpubicAngle { get; set; }
-        public long? SciaticNotch { get; set; }
-        public long? PubicBone { get; set; }
-        public long? PreaurSulcus { get; set; }
-        public long? MedialIpRamus { get; set; }
-        public long? DorsalPitting { get; set; }
+        public int? VentralArc { get; set; }
+        public int? SubpubicAngle { get; set; }
+        public int? SciaticNotch { get; set; }
+        public int? PubicBone { get; set; }
+        public int? PreaurSulcus { get; set; }
+        public int? MedialIpRamus { get; set; }
+        public int? DorsalPitting { get; set; }
         public string ForemanMagnum { get; set; }
         public double? FemurHead { get; set; }
         public double? HumerusHead { get; set; }
@@ -45,13 +45,13 @@ namespace Fag_el_Gamous.Models
         public double? FemurLength { get; set; }
         public double? HumerusLength { get; set; }
         public double? TibiaLength { get; set; }
-        public long? Robust { get; set; }
-        public long? SupraorbitalRidges { get; set; }
-        public long? OrbitEdge { get; set; }
-        public long? ParietalBossing { get; set; }
-        public long? Gonian { get; set; }
-        public long? NuchalCrest { get; set; }
-        public long? ZygomaticCrest { get; set; }
+        public int? Robust { get; set; }
+        public int? SupraorbitalRidges { get; set; }
+        public int? OrbitEdge { get; set; }
+        public int? ParietalBossing { get; set; }
+        public int? Gonian { get; set; }
+        public int? NuchalCrest { get; set; }
+        public int? ZygomaticCrest { get; set; }
         public string CranialSuture { get; set; }
         public double? MaximumCranialLength { get; set; }
         public double? MaximumCranialBreadth { get; set; }
@@ -65,22 +65,22 @@ namespace Fag_el_Gamous.Models
         public string ArtifactsDescription { get; set; }
         public string HairColor { get; set; }
         public string PreservationIndex { get; set; }
-        public long? HairTaken { get; set; }
-        public long? SoftTissueTaken { get; set; }
-        public long? BoneTaken { get; set; }
-        public long? ToothTaken { get; set; }
-        public long? TextileTaken { get; set; }
+        public int? HairTaken { get; set; }
+        public int? SoftTissueTaken { get; set; }
+        public int? BoneTaken { get; set; }
+        public int? ToothTaken { get; set; }
+        public int? TextileTaken { get; set; }
         public string DescriptionOfTaken { get; set; }
-        public long? ArtifactFound { get; set; }
+        public int? ArtifactFound { get; set; }
         public double? EstimateAge { get; set; }
         public double? EstimateLivingStature { get; set; }
         public string ToothAttrition { get; set; }
         public string ToothEruption { get; set; }
         public string PathologyAnomalies { get; set; }
         public string EpiphysealUnion { get; set; }
-        public long? YearFound { get; set; }
+        public int? YearFound { get; set; }
         public string MonthFound { get; set; }
-        public long? DayFound { get; set; }
+        public int? DayFound { get; set; }
         public string HeadDirection { get; set; }
     }
 }
